requirements.txt
torch>=2.2.0
torchvision>=0.17.0
flask>=3.0.0
numpy>=1.26.0
pandas>=2.1.0
requests>=2.31.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
dash>=2.15.0
plotly>=5.18.0
aiohttp>=3.9.0
nest-asyncio>=1.6.0
